fixing fseek conversion warning
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk/ac@3776 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ source/src/oggstream.cpp
@@ -9,7 +9,7 @@
 
 static int oggseek(FILE *f, ogg_int64_t off, int whence)
 {
-    return f ? fseek(f, off, whence) : -1;
+    return f ? fseek(f, (long)off, whence) : -1;
 }
 
 ov_callbacks oggcallbacks =
